-Restored Cmake directory settings to build into the ${WMV_BASE_PATH}/bin at all times. -Continued 64-bit code updates -Installed Qt MainWindow & primary widgets. (Most functionality NYI) -Installed Qt Splash Screen -Updated Canvas Sizes -Canvas sizes now take (some) docked widgets' sizes into account. (Undock widgets to bypass this) -Fixed some CMake issues with ListFileGenerator & UpdateManager. -Filled in enUS translation file. (With that we have so far...) Affected Issues: WMV-297 WMV-363 WMV-364 WMV-366 WMV-371

+++ src/games/wow/CASCFile.cpp
@@ -112,7 +112,7 @@ bool CASCFile::isAlreadyOpened()
     return false;
 }
 
-bool CASCFile::getFileSize(unsigned int & s)
+bool CASCFile::getFileSize(size_t & s)
 {
   bool result = false;
