+++ src/games/wow/HardDriveFile.h
@@ -31,7 +31,7 @@ class _HARDDRIVEFILE_API_ HardDriveFile : public CASCFile
   protected:
     virtual bool openFile();
     virtual bool isAlreadyOpened();
-    virtual bool getFileSize(unsigned int & s);
+    virtual bool getFileSize(size_t & s);
     virtual unsigned long readFile();
     virtual bool doPostCloseOperation();
